+++ sentry/src/main/java/io/sentry/SentryClient.java
@@ -506,6 +506,10 @@ public final class SentryClient implements ISentryClient {
                 e,
                 "An exception occurred while processing event by processor: %s",
                 processor.getClass().getName());
+        options
+            .getClientReportRecorder()
+            .recordLostEvent(DiscardReason.CALLBACK_ERROR, DataCategory.Error);
+        return null;
       }
 
       if (event == null) {
@@ -557,6 +561,8 @@ public final class SentryClient implements ISentryClient {
                 e,
                 "An exception occurred while processing log event by processor: %s",
                 processor.getClass().getName());
+        recordLostLogEvent(DiscardReason.CALLBACK_ERROR, eventBeforeProcessor);
+        return null;
       }
 
       if (event == null) {
@@ -590,6 +596,8 @@ public final class SentryClient implements ISentryClient {
                 e,
                 "An exception occurred while processing metrics event by processor: %s",
                 processor.getClass().getName());
+        recordLostMetricsEvent(DiscardReason.CALLBACK_ERROR, eventBeforeProcessor);
+        return null;
       }
 
       if (event == null) {
@@ -622,6 +630,14 @@ public final class SentryClient implements ISentryClient {
                 e,
                 "An exception occurred while processing transaction by processor: %s",
                 processor.getClass().getName());
+        options
+            .getClientReportRecorder()
+            .recordLostEvent(DiscardReason.CALLBACK_ERROR, DataCategory.Transaction);
+        options
+            .getClientReportRecorder()
+            .recordLostEvent(
+                DiscardReason.CALLBACK_ERROR, DataCategory.Span, spanCountBeforeProcessor + 1);
+        return null;
       }
       final int spanCountAfterProcessor = transaction == null ? 0 : transaction.getSpans().size();
 
@@ -675,6 +691,10 @@ public final class SentryClient implements ISentryClient {
                 e,
                 "An exception occurred while processing replay event by processor: %s",
                 processor.getClass().getName());
+        options
+            .getClientReportRecorder()
+            .recordLostEvent(DiscardReason.CALLBACK_ERROR, DataCategory.Replay);
+        return null;
       }
 
       if (replayEvent == null) {
@@ -709,6 +729,10 @@ public final class SentryClient implements ISentryClient {
                 e,
                 "An exception occurred while processing feedback event by processor: %s",
                 processor.getClass().getName());
+        options
+            .getClientReportRecorder()
+            .recordLostEvent(DiscardReason.CALLBACK_ERROR, DataCategory.Feedback);
+        return null;
       }
 
       if (feedbackEvent == null) {